Fix external bootstraper
